# aacheng_bazaar__r2-presign-attach-35952bf-v1/environment/repo/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Bazaar API
  version: 0.3.0
  description: |
    Bazaar — the embeddable P2P marketplace primitive. Headless, multi-tenant
    commerce API: Listings and Reviews. Integrators own identity and UX;
    Bazaar owns inventory, reputation, and media.

    The Contact module is descoped from the MVP. Buyer→seller contact is
    expressed through the `actions` enum on listings (`contact`); the host
    app owns the actual contact handoff via its own native channel. Bazaar
    stores no contact handles, threads, or offers.

    # Conventions
    - **Auth**: every `/v1` request carries `X-Bazaar-Key` (key_id),
      `X-Bazaar-Timestamp` (unix seconds), and `X-Bazaar-Signature`
      (HMAC-SHA256 over method + path + raw query + sha256(body) + timestamp,
      300s window, single-use nonce). Auth resolves the key to exactly one
      `app_id`; all data is tenant-isolated by `app_id` (RLS backstop).
      Exempt: `/healthz`, tenant signup (`POST /v1/apps`), and the first key
      issuance (which uses a one-time `provisioning_token` instead).
    - **Who is acting**: auth proves which *app* is calling, not which user.
      Seller-gated writes therefore carry `acting_user_id` in the body —
      never a header, so it stays inside the HMAC-signed body hash — checked
      against the listing's `seller_user_id` (403 `seller_only` on mismatch).
      The `actions` list is a rendering hint; this check is the real
...
      a prior read showed `edit` in `actions`.
...
paths:
...
  /v1/reviews:
    post:
      summary: Create a review
      description: |
        Eligibility is gated on the verified interaction:

        - the author must be the recorded buyer of the referenced listing
          (`listing.buyer_user_id == author_user_id`) → otherwise 403
          `review_not_eligible`.
        - `subject_user_id` (who is reviewed) is **derived server-side** from
          the listing's `seller_user_id` — the request MUST NOT supply a subject;
          the server ignoring a client-supplied subject is how the aggregate
          (the one trust signal in MVP) stays uncorruptible. A verified buyer
          posting against any subject would poison reputation.
        - One review per author per listing
          (`UNIQUE(app_id, author_user_id, listing_id) NULLS NOT DISTINCT`) →
          duplicate → 409 `review_exists`.

        On write, the derived subject's aggregate is updated
        (Postgres-authoritative, Redis write-through). Owner: S2.
      operationId: createReview
      tags: [reviews]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewCreateRequest" }
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        default: { $ref: "#/components/responses/Error" }
    get:
      summary: List reviews for a subject (seller history)
      description: "Offset-paginated review history. Owner: P1."
      operationId: listReviews
      tags: [reviews]
      parameters:
        - $ref: "#/components/parameters/SubjectUserId"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page of reviews
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewPage" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/Error" }
  /v1/reviews/aggregate:
    get:
      summary: Aggregate reputation for a subject
      description: |
        Postgres-authoritative aggregate with Redis write-through cache and a
        rebuild path (a Redis-only store has no rebuild — one eviction would
        silently corrupt seller reputation). `avg_rating` is **null, not 0**,
        for unrated sellers — avoids rendering a "0-star" seller. Owner: P1.

        Note: this resolves the README's `GET /v1/users/{id}/reputation`
        sketch — the endpoint lives here so it stays under the `/reviews`
        router prefix.
      operationId: getReviewAggregate
      tags: [reviews]
      parameters:
        - $ref: "#/components/parameters/SubjectUserId"
      responses:
        "200":
          description: Aggregate reputation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewAggregate" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    BazaarHMAC:
      type: apiKey
      in: header
      name: X-Bazaar-Key
      description: |
        API key id. Full auth also requires `X-Bazaar-Timestamp` and
        `X-Bazaar-Signature` (HMAC-SHA256 over method + path + raw query +
        sha256(body) + timestamp, 300s window, single-use nonce in Redis).
    ProvisioningBearer:
      type: http
      scheme: bearer
      bearerFormat: provisioning_token
      description: |
        One-time provisioning_token from POST /v1/apps (first key only, 15-min TTL).
        Used as `Authorization: Bearer <token>` to issue the first API key before
        the integrator has any HMAC secret. Subsequent rotations use BazaarHMAC.
        Server-side SDK path must never expose this token to browsers — it is a
        secret that returns exactly once.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: UUID; retries with the same key never double-create.
      schema: { type: string, format: uuid }
...
    SubjectUserId:
      name: subject_user_id
      in: query
      required: true
      description: Opaque integrator user id being reviewed.
      schema: { type: string }
    Offset:
      name: offset
      in: query
      required: false
      description: Results to skip (default 0). An invalid offset → 400, not 500.
      schema: { type: integer, default: 0, minimum: 0 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: |
        Authentication failure — one of six granular codes (deliberately
        granular, decided Aug 4, so integrators can distinguish clock skew
        from signing bugs from replay): `missing_auth_headers` (any of
        X-Bazaar-Key / X-Bazaar-Timestamp / X-Bazaar-Signature absent),
        `unknown_key` (key_id does not resolve to an active key),
        `malformed_timestamp` (not unix seconds), `stale_timestamp`
        (outside the 300s window), `invalid_signature` (HMAC does not
        verify over method + path + query + body hash + timestamp),
        `replay_detected` (nonce already used — single-use store in Redis).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error:
              code: invalid_signature
              message: HMAC signature invalid or timestamp outside 300s window.
              request_id: req_7f2a9c
    Forbidden:
      description: Authenticated but not permitted (e.g. review_not_eligible)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error:
              code: review_not_eligible
              message: Author is not the verified buyer of this listing.
              request_id: req_2c41b7
...
    ValidationFailed:
      description: |
        Bad input, missing lat/lng pair, invalid offset, or failed image
        verification.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error:
              code: validation_failed
              message: lat and lng are required together for geo filtering.
              request_id: req_31dd08
    Conflict:
      description: |
        One review per author per listing (`review_exists`). On requests
        sending `Idempotency-Key`, a 409 can also carry `request_in_flight`
        or `idempotency_mismatch` — see IdempotencyConflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error:
              code: review_exists
              message: Author has already reviewed this listing.
              request_id: req_44de1a
...
  schemas:
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Stable machine-readable code.
          example: listing_not_found
        message: { type: string }
        request_id: { type: string, example: req_9f1c2d }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/Error" }
    Pagination:
      type: object
      description: |
        Offset pagination envelope. `total` is the full matching count for the
        filters; `has_more` is `offset + limit < total`.
      required: [total, limit, offset, has_more]
      properties:
        total: { type: integer, minimum: 0, example: 142 }
        limit: { type: integer, minimum: 1, example: 20 }
        offset: { type: integer, minimum: 0, example: 0 }
        has_more: { type: boolean, example: true }
...
    ReviewCreateRequest:
      type: object
      description: |
        `subject_user_id` is NOT accepted from the client — it is derived
        server-side from `listing.seller_user_id`. This closes the trust-loop
        hole where a verified buyer could post against any subject.
      required: [author_user_id, listing_id, rating]
      properties:
        author_user_id:
          type: string
          description: Must equal the listing's recorded buyer_user_id.
        listing_id:
          type: string
          description: Prefixed listing id (lst_<uuid>) the review is anchored to.
        rating: { type: integer, minimum: 1, maximum: 5 }
        body: { type: string }
    Review:
      type: object
      required: [id, app_id, subject_user_id, author_user_id, listing_id, rating, created_at]
      description: |
        MVP reviews are always listing-anchored (listing_id required). The DB
        column is nullable for the UNIQUE ... NULLS NOT DISTINCT constraint,
        but the API shape is non-null.
      properties:
        id:
          type: string
          description: Prefixed public id (rev_<uuid>); DB stores a bare uuid.
          example: rev_b2c3d4e5
        app_id: { type: string, format: uuid }
        subject_user_id:
          type: string
          description: Derived server-side from the listing's seller_user_id (not client-supplied).
        author_user_id: { type: string }
        listing_id:
          type: string
          description: Prefixed listing id (lst_<uuid>) the review is anchored to; empty when unanchored is v1.
        rating: { type: integer, minimum: 1, maximum: 5 }
        body: { type: string }
        created_at: { type: string, format: date-time }
    ReviewPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Review" }
        pagination: { $ref: "#/components/schemas/Pagination" }
    ReviewAggregate:
      type: object
      required: [subject_user_id, avg_rating, review_count, rating_counts]
      properties:
        subject_user_id: { type: string }
        avg_rating:
          type: [number, "null"]
          description: 1 decimal; null (not 0) when review_count == 0.
          example: 4.8
        review_count: { type: integer, example: 23 }
        rating_counts:
          type: object
          description: Histogram keyed by rating ("1".."5").
          additionalProperties: { type: integer }
          example: { "1": 0, "2": 0, "3": 1, "4": 4, "5": 18 }
security:
  - BazaarHMAC: []